+++ lib/main_canvas/grid_custom_painter.dart
@@ -1,7 +1,6 @@
 // ignore_for_file: unnecessary_this
 
 import 'package:diagrams/flow_elements/abstract_flow_element.dart';
-import 'package:diagrams/flow_elements/bloc/arrows/arrow_model.dart';
 import 'package:diagrams/flow_elements/bloc/arrows/draw_arrows_bloc.dart';
 import 'package:diagrams/flow_elements/bloc/arrows/draw_arrows_event.dart';
 import 'package:flutter/material.dart';
@@ -61,8 +60,11 @@ class GridCustomPainter extends CustomPainter {
     void onPanDown(DragDownDetails details) {
       var newPoint = normalizedPointToGrid(details.localPosition);
       var elementAnchorPointFound = flowElementsList?.firstWhereOrNull((e) =>
-          e.anchorPointsModelMap!.anchorPointList.firstWhereOrNull(
-              (e) => e.anchorPointPositionRelativeToParent == newPoint) !=
+          e.anchorPointsModelMap!.anchorPointList.firstWhereOrNull((e) {
+            return (e.anchorPointPositionRelativeToParent - newPoint)
+                    .distanceSquared <=
+                300;
+          }) !=
           null);
       if (elementAnchorPointFound == null) return;
 
@@ -174,25 +176,4 @@ class GridCustomPainter extends CustomPainter {
   bool shouldRepaint(covariant CustomPainter oldDelegate) {
     return false;
   }
-}
-
-class ArrowCustomPainter extends CustomPainter {
-  final List<ArrowModel> arrowModelList;
-
-  ArrowCustomPainter({required this.arrowModelList});
-  @override
-  void paint(Canvas canvas, Size size) {
-    var pointPaint = Paint()
-      ..color = Colors.black
-      ..strokeWidth = 2
-      ..style = PaintingStyle.stroke
-      ..strokeCap = StrokeCap.square;
-
-    for (var arrow in arrowModelList) {
-      canvas.drawPath(arrow.arrowPath, pointPaint);
-    }
-  }
-
-  @override
-  bool shouldRepaint(covariant CustomPainter oldDelegate) => true;
 }
